requirements.txt
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
ratelimit==2.2.1
python-telegram-bot==21.0.1
aiohttp==3.8.6
